+++ CHANGELOG.md
@@ -1,5 +1,13 @@
 # Change Log
 
+## 0.0.3-1 - 2017-06-09
+
+* fork from 0.0.3
+* Upgrade fluency from 0.0.10 to 1.3.0
+* Upgrade jackson from 2.6.1 to 2.6.7
+* Support batchSize with default value 1000
+* Remove transaction rollback when no event can be retrieved
+
 ## 0.0.3 - 2016-5-12
 
 * Support buffered events storing local file mechanism